Updated ForceKeepFlags to keep a count of each flag Extracted more methods into partial classes of Vector
