+++ .editorconfig
@@ -1,0 +1,4 @@
+﻿[*.cs]
+
+# CS0162: Unreachable code detected
+dotnet_diagnostic.CS0162.severity = warning
